% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consol_functions.R
\name{addl_expt_attribute}
\alias{addl_expt_attribute}
\title{Create a new column with additional experimental information.}
\usage{
addl_expt_attribute(table, column, contents)
}
\arguments{
\item{table}{A tibble}

\item{column}{A character, the new column name}

\item{contents}{A character, the new column contents}
}
\value{
A tibble with the appended column
}
\description{
Create a new column with additional experimental information.
}
\examples{
mod_iris <- addl_expt_attribute(dplyr::as_tibble(iris), "flower_type", "iris")
}
