% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{are_colors}
\alias{are_colors}
\title{function to check that colors are viable}
\usage{
are_colors(x)
}
\arguments{
\item{x}{list of colors}
}
\value{
list of T/F, TRUE if a real color and False if not
}
\description{
function to check that colors are viable
}
