% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_functions.R
\name{calc_fold_change}
\alias{calc_fold_change}
\title{Calculates the fold change using deltadeltaCt}
\usage{
calc_fold_change(
  table,
  house_gene,
  control,
  expt_conditions = c(""),
  sep = "",
  full_table = "no"
)
}
\arguments{
\item{table}{a tibble with columns sample_name, target_name, and ct}

\item{house_gene}{a character, the target_name which is the housekeeping gene}

\item{control}{a character or character list, conditions which represent the control condition}

\item{expt_conditions}{a character list, conditions which make up sample_name, will become titles of new columns}

\item{sep}{a character, the separation character used in sample_name to separate conditions}

\item{full_table}{a character, "yes" or "no", if yes, will return all intermediate statistics}
}
\value{
a tibble with fold changes calculated based on house_gene and control conditions
}
\description{
Calculates the fold change using deltadeltaCt
}
