% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_color}
\alias{check_color}
\title{checkmate function to check for proper colors}
\usage{
check_color(table, color, color_choices)
}
\arguments{
\item{table}{a tibble}

\item{color}{a character list, a list of columns which to assign the colors by}

\item{color_choices}{a character list, list of the colors}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check for proper colors
}
