% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_ct_numeric}
\alias{check_ct_numeric}
\title{checkmate function to check the ct column has proper numeric format}
\usage{
check_ct_numeric(table, ct)
}
\arguments{
\item{table}{a tibble}

\item{ct}{a character, the column title ct}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check the ct column has proper numeric format
}
