% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_house_gene}
\alias{check_house_gene}
\title{checkmate function to check the housekeeping gene}
\usage{
check_house_gene(table, house_gene)
}
\arguments{
\item{table}{a tibble}

\item{house_gene}{a character, the house_keeping gene}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check the housekeeping gene
}
