% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_not_col_name}
\alias{check_not_col_name}
\title{checkmate function to check that a column name is not already in the column}
\usage{
check_not_col_name(table, column)
}
\arguments{
\item{table}{a tibble}

\item{column}{a character, the new column title}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check that a column name is not already in the column
}
