% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_replicates}
\alias{check_replicates}
\title{checkmate function to check for the replicates}
\usage{
check_replicates(table, replicates)
}
\arguments{
\item{table}{a tibble}

\item{replicates}{a character, tests for the replicates being in table from multiple expts}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check for the replicates
}
