% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_sample_control}
\alias{check_sample_control}
\title{checkmate function to check for control condition in the table}
\usage{
check_sample_control(table, sample_control)
}
\arguments{
\item{table}{a tibble}

\item{sample_control}{a string, made from inputs to calc_fold_change}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check for control condition in the table
}
