% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_sep}
\alias{check_sep}
\title{checkmate function to check for the separater}
\usage{
check_sep(table, expt_conditions = c(""), sep = "")
}
\arguments{
\item{table}{a tibble}

\item{expt_conditions}{a character list, a list of columns to combine}

\item{sep}{a character, the separator used to combine all the columns}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check for the separater
}
