% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_functions.R
\name{clean_qPCR_files_check_NTC}
\alias{clean_qPCR_files_check_NTC}
\title{Read in qPCR files, clean them and report on controls}
\usage{
clean_qPCR_files_check_NTC(file_location, file_name, undetermined = "set40")
}
\arguments{
\item{file_location}{a character, path to the file, name not included, must have trailing /}

\item{file_name}{a character, file name with extension}

\item{undetermined}{a character, "set40" or "remove"; set40 takes undetermined values and sets them to 40, otherwise they are removed. default is set40}
}
\value{
a tibble with cleaned data, plus printed output on controls
}
\description{
Read in qPCR files, clean them and report on controls
}
