% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_calculations.R
\name{conc_calc}
\alias{conc_calc}
\title{Summarize data and generate linear regression to fit standards, report linear reg variables
and the plot of the summarized data}
\usage{
conc_calc(data_table, yint, slope, avg_length, dilution_fact, kit)
}
\arguments{
\item{data_table}{a tibble with unknown samples, 3 columns: sample_name, target_name, and ct}

\item{yint}{a numeric, the y intercept of the regression line}

\item{slope}{a numeric, the slope of the regression line}

\item{avg_length}{a numeric, the avg length of the molecules in the library}

\item{dilution_fact}{a numeric, the dilution factor used for quantification ie for 1:1000, you put 1000}

\item{kit}{a string either KAPA or NEB}
}
\value{
a tibble with original library concentrations
}
\description{
Summarize data and generate linear regression to fit standards, report linear reg variables
and the plot of the summarized data
}
