% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_calculations.R
\name{linear_reg}
\alias{linear_reg}
\title{Summarize data and generate linear regression to fit standards, report linear reg variables
and the plot of the summarized data}
\usage{
linear_reg(table)
}
\arguments{
\item{table}{a tibble with quantities and corresponding ct values}
}
\value{
list: (yint, slope, graph of standards)
}
\description{
Summarize data and generate linear regression to fit standards, report linear reg variables
and the plot of the summarized data
}
