% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{qPCR_graph}
\alias{qPCR_graph}
\title{A function to graph fold change qPCR data}
\usage{
qPCR_graph(
  table,
  x = "target_name",
  y = "fold_change",
  house_gene,
  color = "",
  xlab = "",
  ylab = "Fold Change\\n",
  colorlab = "Expt",
  color_choices = c("#336699", "#9999FF", "black", "red"),
  ycoord_max = 5,
  replicates = "no_replicates",
  bars = "std_error",
  hline = "yes",
  coord_cart = "yes",
  theme = "jaclyn"
)
}
\arguments{
\item{table}{a tibble, columns are expt conditions, fold change, error, target_name}

\item{x}{a character, the column title for the variable to be plotted on the x axis, target_name}

\item{y}{a character, the column title for the variable to be plotted on the y axis, fold_change}

\item{house_gene}{a character, gene to use as the housekeeping gene}

\item{color}{a character/character list, the columns to group into color}

\item{xlab}{a character, the label for the x axis, default is ""}

\item{ylab}{a character, the label for the y axis, default is "Fold Change\\n"}

\item{colorlab}{a character, the label for the legend based on color, default is "Expt"}

\item{color_choices, }{a character list, the colors you want to use for the graph}

\item{ycoord_max}{a numeric, the max value for the fold change displayed on the graph}

\item{replicates}{a character, "no replicates" or the list of columns to use for the replicates}

\item{bars}{a character: "std_dev" or "std_error" type of error bars you will get on your graph}

\item{hline}{a character: "yes" or "no", if yes, a horizontal line for fold change is 1 is added}

\item{coord_cart}{a character: "yes" or "no", if no graph scaled to extent of data}

\item{theme}{a character: "jaclyn" or "default", if default, the original theme applied and you can add your own theme later}
}
\value{
summary statistics table is printed and graph object is returned
}
\description{
A function to graph fold change qPCR data
}
