% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCRr.R
\docType{package}
\name{qPCRr}
\alias{qPCRr}
\title{qPCRr: a package for cleaning, analyzing, and graphing qPCR data.}
\description{
The qPCRr package provides many functions.
}
\section{qPCRr functions}{

addl_expt_attribute, calc_fold_change, clean_qPCR_files_check_NTC, combine_expts, qPCR_graph, reformat_table, reorder_samples, conc_calc, linear_reg, remove_outliers, std_curve_file_clean
}

