% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_functions.R
\name{reformat_table}
\alias{reformat_table}
\title{reformat other data tables into one which can be used by calc_fold_change}
\usage{
reformat_table(table, sample, target, ct, sep = "_")
}
\arguments{
\item{table}{a data frame or tibble}

\item{sample}{a character, the column, or list of columns which make up sample_name}

\item{target}{a character, the column which makes up the targeted genes}

\item{ct}{a character, the column which makes up the ct values of entries}

\item{sep}{a character, the character used to combine columns together}
}
\value{
a tibble in the proper format for calc_fold_change
}
\description{
reformat other data tables into one which can be used by calc_fold_change
}
