% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_fun_std_curve.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers greater than 0.2 ct values from the triplicate}
\usage{
remove_outliers(table)
}
\arguments{
\item{table}{a tibble with quantities and corresponding ct values}
}
\value{
a tibble; if range is less than 0.2 return all, drop outliers
}
\description{
Remove outliers greater than 0.2 ct values from the triplicate
}
