% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consol_functions.R
\name{reorder_samples}
\alias{reorder_samples}
\title{Order the entries in a column so that they are displayed in this order for graphs, rather than alphabetically.}
\usage{
reorder_samples(table, column, list)
}
\arguments{
\item{table}{A tibble}

\item{column}{A character, the title of a column of table to be ordered}

\item{list}{A character list, a list of column entries in the correct order for graphing}
}
\value{
Table is returned with a reordered column stored as a type, factor.
}
\description{
Order the entries in a column so that they are displayed in this order for graphs, rather than alphabetically.
}
\note{
list is a set of characters in the order they should appear in graphs. All entires in the list must be unique and all values must appear in the list.
}
