% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{sum_tbl_calc}
\alias{sum_tbl_calc}
\title{A function to calculate summary statistics}
\usage{
sum_tbl_calc(table, color, house_gene, bars, replicates)
}
\arguments{
\item{table}{a tibble}

\item{color}{a character or character list, a list of columns to use for color}

\item{house_gene}{a character, gene to use as the housekeeping gene}

\item{bars}{a character: "std_dev" or "std_error" type of error bars you will get on your graph}

\item{replicates}{a character, "no replicates" or the list of columns to use for the replicates}
}
\value{
summary statistics table
}
\description{
A function to calculate summary statistics
}
