% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_helper_functions.R
\name{check_sample_control}
\alias{check_sample_control}
\title{checkmate function to check for control condition in the table}
\usage{
check_sample_control(table, control, sep)
}
\arguments{
\item{table}{a tibble}

\item{control}{a character or character list, factors which make up the control condition}

\item{sep}{a character, the separator used to combine the control conditions}
}
\value{
true if ok, an error message if false
}
\description{
checkmate function to check for control condition in the table
}
